requirements.txt
langchain
langchain-openai
langchain-mongodb
langgraph
uvicorn
pymongo
pandas
python-dotenv
requests
fastapi
langchain-community
python-multipart
openpyxl
